% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-prospect-sequences.R
\name{add_prospect_seq}
\alias{add_prospect_seq}
\title{Add prospect to a sequence}
\usage{
add_prospect_seq(prospect_id, sequence_id, mailbox_id)
}
\arguments{
\item{prospect_id}{The numeric ID for an Outreach prospect.}

\item{sequence_id}{The numeric ID for an Outreach sequence.}
}
\description{
Adds a prospect to a sequence given a prospect ID and a sequence ID.
}
