% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailboxes.R
\name{get_mailboxes}
\alias{get_mailboxes}
\title{Get Outreach Mailboxes}
\usage{
get_mailboxes(token = Sys.getenv("OUTREACH_TOKEN"))
}
\arguments{
\item{token}{Your Outreach.io token which should be stored in the environment variable \code{OUTEREACH_TOKEN} if authentication was done with \code{outreach_auth()}}
}
\value{
A tibble containing Outreach users and their respective email address, signatures, and \code{user_id}—a.k.a. their mailbox ID.
}
\description{
\code{get_mailboxes()} returns mailbox information for Outreach users. The "mailbox" is a reference to the user and user email address from which sequences and emails will be sent from or associated with. The mailbox identifier is referred to as the \code{user_id} in the Outreach API.
}
\examples{

\dontrun{
all_mailboxes <- get_mailboxes()
}

}
