% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{get_sequences}
\alias{get_sequences}
\title{Get Outreach Sequence Information}
\usage{
get_sequences(token = Sys.getenv("OUTREACH_TOKEN"))
}
\arguments{
\item{token}{Your Outreach.io token which should be stored in the environment variable \code{OUTEREACH_TOKEN} if authentication was done with \code{outreach_auth()}}
}
\value{
A tibble containing sequence identifiers and associated information.
}
\description{
Returns a tibble containing sequence ID, attributes, and relationships. The sequence ID is used to add prospects to sequences using \code{add_prospect_seq()}.
}
