% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{outreach_auth}
\alias{outreach_auth}
\title{Authenticate an Outreach application}
\usage{
outreach_auth(key, secret, redirect, scopes)
}
\arguments{
\item{key}{The application or client ID provided to you by Outreach.io.}

\item{secret}{The secret provided by Outreach.io.}

\item{redirect}{The callback URI or URL provided by Outreach.io.}

\item{scopes}{A character scalar containing the API privileges to be granted. Each scope is to be separated by a space. Each scope is to be defined by the name of the resource and the access level, e.g. \code{"prospects.read sequenceStates.all"}.}
}
\description{
Authenticate an Outreach application
}
