% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{refresh_auth_token}
\alias{refresh_auth_token}
\title{Refresh Outreach Authentication Token}
\usage{
refresh_auth_token(
  key,
  secret,
  redirect,
  refresh_token = Sys.getenv("OUTREACH_REFRESH")
)
}
\arguments{
\item{key}{The application or client ID provided to you by Outreach.io.}

\item{secret}{The secret provided by Outreach.io.}

\item{redirect}{The callback URI or URL provided by Outreach.io.}
}
\description{
Authentication tokens granted by the Outreach.io API last for only 120 minutes. After which period the tokens need to be refreshed using the refresh token provided alongside the grant token.
}
