% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-prospects.R
\name{search_prospects}
\alias{search_prospects}
\title{Search Outreach.io prospects.}
\usage{
search_prospects(
  owner_email,
  owner_id,
  account_name,
  tag,
  token = Sys.getenv("OUTREACH_TOKEN")
)
}
\arguments{
\item{owner_email}{Prospect owner email address(es). Optional.}

\item{owner_id}{Prospect owner ID(s) (also known as mailbox ID). Optional.}

\item{account_name}{Account name(s). Optional.}

\item{tag}{Prospect tag(s). Optional.}

\item{token}{Your Outreach.io token which should be stored in the environment variable \code{OUTEREACH_TOKEN} if authentication was done with \code{outreach_auth()}}
}
\description{
\code{search_prospects()} allows you to query the Outreach API for prospects (up to 1000 as of right now) based on the owner ID or email, account name, or tag. For each of these arguments you can provide a vector of length > 0. If you provide, for example, a vector with two tags it will return prospects that match at least one of the conditions.
}
\examples{
search_prospects(owner_email = "josiah@rstudio.com")
search_prospect(tag = c("my-favorite-prospects", "okay-prospects"))

}
