% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-prospects.R
\name{search_prospects}
\alias{search_prospects}
\title{Search Outreach.io prospects.}
\usage{
search_prospects(
  owner_email,
  owner_id,
  account_name,
  tag,
  token = Sys.getenv("OUTREACH_TOKEN")
)
}
\arguments{
\item{owner_email}{Prospect owner email address(es). Optional.}

\item{owner_id}{Prospect owner ID(s) (also known as mailbox ID). Optional.}

\item{account_name}{Account name(s). Optional.}

\item{tag}{Prospect tag(s). Optional.}

\item{token}{Your Outreach.io token which should be stored in the environment variable \code{OUTEREACH_TOKEN} if authentication was done with \code{outreach_auth()}}
}
\description{
Search Outreach.io prospects.
}
